Exit script if fetch ip failed

+++ get_ip.sh
@@ -11,10 +11,14 @@ done
 ## get ip now
 
 curl http://members.3322.org/dyndns/getip > $ip_now
+## exit if fetch ip failed
+if [ ! -s $ip_now ]; then
+        exit 0
+fi
 
 /usr/bin/diff $ip_now $ip_old
 
-if [ $? != 0 ];then
+if [ $? != 0 ]; then
         sed -i "\$a$(date) ------ IP change detected, previous ip: $(cat $ip_old), new ip: $(cat $ip_now)." $log
         cat $ip_now > $ip_old
         curl "https://sc.ftqq.com/SCUKEY.send?text=$(cat $ip_now)";
